+++ src/main/kotlin/org/ender_development/template_mod/TemplateModReference.kt
@@ -11,7 +11,7 @@ object TemplateModReference {
     const val GROOVY_VERSION = "4.0.26"
     const val CATALYX_VERSION = "0.1.0"
     const val ASSETMOVER_VERSION = 2.5
-    const val GROOVYSCRIPT_VERSION = "1.3.3"
+    const val GROOVYSCRIPT_VERSION = "1.3.4"
     const val MOD_NAME = "Template Mod"
     const val CRAFTTWEAKER_VERSION = "4.1.20.711"
     const val KOTLIN_LANGUAGE_ADAPTER = "io.github.chaosunity.forgelin.KotlinAdapter"
@@ -23,6 +23,6 @@ object TemplateModReference {
     const val MODULARUI_VERSION = "3.0.8"
     const val FORGELIN_VERSION = "2.3.0.0"
     const val VERSION = "0.1.0"
-    const val HADENOUGHITEMS_VERSION = "4.29.14"
+    const val HADENOUGHITEMS_VERSION = "4.29.15"
     const val MODID = "template_mod"
 }